Review Changes: remove useless variable fontList + correct return comment
